+++ cli/tests/cli.rs
@@ -535,7 +535,6 @@ fn different_supervisor_modes_in_public_configs() -> Result<(), failure::Error> 
         database: Default::default(),
         thread_pool_size: None,
         connect_list: Default::default(),
-        keys: Default::default(),
     };
 
     let testnet_dir = tempfile::tempdir()?;